+++ lib/src/widget/widget.dart
@@ -3,3 +3,4 @@ export 'controller.dart';
 export 'dot_control.dart';
 export 'edge_alignment.dart';
 export 'crop_result.dart';
+export 'initial_rect_builder.dart';