+++ test/logic/parser/image_image_parser_test.dart
@@ -52,7 +52,7 @@ void main() {
             testImage,
             inputFormat: ImageFormat.jpeg,
           ),
-          throwsA(const TypeMatcher<InvalidInputFormatError>()),
+          throwsA(const TypeMatcher<InvalidInputFormatException>()),
         );
       },
     );